Pr-tilbakemelding: Flytt logikk til avstemmingsservice
